pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tictoc-factory"
version = "0.1.0"
description = "Headless TikTok content factory for Raspberry Pi deployment."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.8,<3",
  "Pillow>=11.0",
  "PyYAML>=6.0",
  "python-dotenv>=1.0",
  "requests>=2.32",
  "typer>=0.12",
  "typing-extensions>=4.12",
]

[project.optional-dependencies]
dev = [
  "pyright>=1.1.390",
  "pytest>=8.3",
  "pytest-cov>=6.0",
  "ruff>=0.8",
]

[project.scripts]
tictoc-factory = "tictoc_factory.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "--cov=src/tictoc_factory --cov-report=term-missing"

[tool.ruff]
line-length = 140
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "SIM", "RUF"]

[tool.pyright]
include = ["src", "tests"]
venvPath = "."
venv = ".venv"
pythonVersion = "3.11"
typeCheckingMode = "standard"
